move boilerplate event stuff
- logging performed in single method of base class
