+++ connector/src/main/java/org/geysermc/connector/network/UpstreamPacketHandler.java
@@ -28,7 +28,7 @@ package org.geysermc.connector.network;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.JsonNodeType;
 import com.nimbusds.jose.JWSObject;
-import com.nukkitx.protocol.bedrock.handler.BedrockPacketHandler;
+import com.nukkitx.protocol.bedrock.BedrockPacket;
 import com.nukkitx.protocol.bedrock.packet.*;
 import com.nukkitx.protocol.bedrock.util.EncryptionUtils;
 import net.minidev.json.JSONObject;
@@ -55,14 +55,15 @@ import java.security.interfaces.ECPublicKey;
 import java.security.spec.ECGenParameterSpec;
 import java.util.UUID;
 
-public class UpstreamPacketHandler implements BedrockPacketHandler {
-
-    private GeyserConnector connector;
-    private GeyserSession session;
+public class UpstreamPacketHandler extends LoggingPacketHandler {
 
     public UpstreamPacketHandler(GeyserConnector connector, GeyserSession session) {
-        this.connector = connector;
-        this.session = session;
+        super(connector, session);
+    }
+
+    private boolean translateAndDefault(BedrockPacket packet) {
+        Registry.BEDROCK.translate(packet.getClass(), packet, session);
+        return defaultHandler(packet);
     }
 
     @Override
@@ -150,188 +151,6 @@ public class UpstreamPacketHandler implements BedrockPacketHandler {
     }
 
     @Override
-    public boolean handle(AdventureSettingsPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(AnimatePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        Registry.BEDROCK.translate(packet.getClass(), packet, session);
-        return true;
-    }
-
-    @Override
-    public boolean handle(BlockEntityDataPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(BlockPickRequestPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(BookEditPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ClientCacheBlobStatusPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ClientCacheMissResponsePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ClientCacheStatusPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ClientToServerHandshakePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(CommandBlockUpdatePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(CommandRequestPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        Registry.BEDROCK.translate(packet.getClass(), packet, session);
-        return true;
-    }
-
-    @Override
-    public boolean handle(ContainerClosePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(CraftingEventPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(EntityEventPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(EntityFallPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(EntityPickRequestPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(EventPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(InteractPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(InventoryContentPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(InventorySlotPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(InventoryTransactionPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ItemFrameDropItemPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(LabTablePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(LecternUpdatePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(LevelEventGenericPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(LevelSoundEventPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(LevelSoundEvent3Packet packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(MapInfoRequestPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(MobArmorEquipmentPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(MobEquipmentPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
     public boolean handle(ModalFormResponsePacket packet) {
         connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
         WindowCache windowCache = session.getWindowCache();
@@ -361,12 +180,6 @@ public class UpstreamPacketHandler implements BedrockPacketHandler {
     }
 
     @Override
-    public boolean handle(MoveEntityAbsolutePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
     public boolean handle(MovePlayerPacket packet) {
         connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
         if (!session.isLoggedIn()) {
@@ -383,595 +196,6 @@ public class UpstreamPacketHandler implements BedrockPacketHandler {
         return false;
     }
 
-    @Override
-    public boolean handle(NetworkStackLatencyPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(PhotoTransferPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(PlayerActionPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(PlayerHotbarPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(PlayerInputPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(PlayerSkinPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(PurchaseReceiptPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(RequestChunkRadiusPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ResourcePackChunkRequestPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(RiderJumpPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ServerSettingsRequestPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SetDefaultGameTypePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SetLocalPlayerAsInitializedPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SetPlayerGameTypePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SubClientLoginPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(TextPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        Registry.BEDROCK.translate(packet.getClass(), packet, session);
-        return true;
-    }
-
-    @Override
-    public boolean handle(AddBehaviorTreePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(AddEntityPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(AddHangingEntityPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(AddItemEntityPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(AddPaintingPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(AddPlayerPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(AvailableCommandsPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(BlockEventPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(BossEventPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(CameraPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ChangeDimensionPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ChunkRadiusUpdatedPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ClientboundMapItemDataPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(CommandOutputPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ContainerOpenPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ContainerSetDataPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(CraftingDataPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(DisconnectPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ExplodePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(LevelChunkPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(GameRulesChangedPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(GuiDataPickItemPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(HurtArmorPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(AutomationClientConnectPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(LevelEventPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(MapCreateLockedCopyPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(MobEffectPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ModalFormRequestPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(MoveEntityDeltaPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(NpcRequestPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(OnScreenTextureAnimationPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(PlayerListPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(PlaySoundPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(PlayStatusPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(RemoveEntityPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(RemoveObjectivePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ResourcePackChunkDataPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ResourcePackDataInfoPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ResourcePacksInfoPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ResourcePackStackPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(RespawnPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ScriptCustomEventPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ServerSettingsResponsePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ServerToClientHandshakePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SetCommandsEnabledPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SetDifficultyPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SetDisplayObjectivePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SetEntityDataPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SetEntityLinkPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SetEntityMotionPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SetHealthPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SetLastHurtByPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SetScoreboardIdentityPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SetScorePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SetSpawnPositionPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SetTimePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SetTitlePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ShowCreditsPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ShowProfilePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(ShowStoreOfferPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SimpleEventPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SpawnExperienceOrbPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(StartGamePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(StopSoundPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(StructureBlockUpdatePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(StructureTemplateDataExportRequestPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(StructureTemplateDataExportResponsePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(TakeItemEntityPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(TransferPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(UpdateAttributesPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(UpdateBlockPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(UpdateBlockPropertiesPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(UpdateBlockSyncedPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(UpdateEquipPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(UpdateSoftEnumPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(UpdateTradePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(AvailableEntityIdentifiersPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(BiomeDefinitionListPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(LevelSoundEvent2Packet packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(NetworkChunkPublisherUpdatePacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(SpawnParticleEffectPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
-    @Override
-    public boolean handle(VideoStreamConnectPacket packet) {
-        connector.getLogger().debug("Handled packet: " + packet.getClass().getSimpleName());
-        return false;
-    }
-
     private void startEncryptionHandshake(PublicKey key) throws Exception {
         KeyPairGenerator generator = KeyPairGenerator.getInstance("EC");
         generator.initialize(new ECGenParameterSpec("secp384r1"));
@@ -985,4 +209,20 @@ public class UpstreamPacketHandler implements BedrockPacketHandler {
         packet.setJwt(EncryptionUtils.createHandshakeJwt(serverKeyPair, token).serialize());
         session.getUpstream().sendPacketImmediately(packet);
     }
+
+    @Override
+    public boolean handle(AnimatePacket packet) {
+        return translateAndDefault(packet);
+    }
+
+    @Override
+    public boolean handle(CommandRequestPacket packet) {
+        return translateAndDefault(packet);
+    }
+
+    @Override
+    public boolean handle(TextPacket packet) {
+        return translateAndDefault(packet);
+    }
+
 }
